Patch up test suite to work with AT-commander changes.

+++ src/tests/platform/timer.cpp
@@ -1,6 +1,6 @@
 #include "util/timer.h"
 
-void openxc::util::time::delayMs(int delayInMs) { }
+void openxc::util::time::delayMs(unsigned long delayInMs) { }
 
 unsigned long openxc::util::time::systemTimeMs() {
     return 0;
